[publishing] 슬라이더 컴포넌트 분리, FeatureSlider로 변경
