+++ src/pages/Home.js
@@ -1,13 +1,5 @@
 import React from "react";
-import Slider from "react-slick";
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
-import "../styles/sliderStyles.css";
-import slider_diary from "../assets/home/slider/slider_diary.png";
-import slider_game from "../assets/home/slider/slider_game.png";
-import slider_gymnastics from "../assets/home/slider/slider_gymnastics.png";
-import slider_dementiacenter from "../assets/home/slider/slider_dementiacenter.png";
-import slider_survey from "../assets/home/slider/slider_survey.png";
+import FeatureSlider from "../component/FeatureSlider";
 import btn_centerMap from "../assets/home/button/btn_centerMap.png";
 import btn_survey from "../assets/home/button/btn_survey.png";
 import btn_diary from "../assets/home/button/btn_diary.png";
@@ -20,7 +12,7 @@ export const HOME_PAGE_PATH = "/home";
 const Home = () => {
   return (
     <>
-      <SimpleSlider />
+      <FeatureSlider />
       <div className="px-5">
         <div className="flex justify-between my-8">
           <MainButton adverb={"간단한"} buttonName={"치매진단"} imgsrc={btn_survey} />
@@ -41,49 +33,11 @@ const Home = () => {
 
 export default Home;
 
-//슬라이더 컴포넌트
-const SimpleSlider = () => {
-  //슬라이더 설정
-  const settings = {
-    dots: true,
-    infinite: true,
-    autoplay: true,
-    arrows: false,
-    autoplaySpeed: 2000,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-  };
-
-  //슬라이더 데이터
-  const sliderDatas = [
-    { imgsrc: slider_diary, sentence1: "일기를 회상하고", sentence2: "나만의 그림을 그려보세요!", color:"bg-[#6100C1]" },
-    { imgsrc: slider_dementiacenter, sentence1: "치매센터가 어디있나요?", sentence2: "가까운 치매센터 찾기", color:"bg-[#356DD9]" },
-    { imgsrc: slider_gymnastics, sentence1: "간단하게 따라할 수 있는", sentence2: "체조들을 만나보세요!", color:"bg-[#FF7D7D]" },
-    { imgsrc: slider_survey, sentence1: "간단한 치매 진단과", sentence2: "이전 결과들을 확인하세요!", color:"bg-[#FF9B52]" },
-    { imgsrc: slider_game, sentence1: "간단한 게임들을 통해", sentence2: "인지력을 강화시켜보세요!", color:"bg-[#FFE68F]" },
-  ];
-
-  return (
-      <Slider {...settings}>
-        {sliderDatas.map((slider, index) => (
-          <div key={index} className={`relative h-64 ${slider.color}`}>
-            <div className={`absolute top-6 sm:top-8 md:top-12 left-6 sm:left-8 md:left-12 text-2xl md:text-3xl font-bold ${index === 4 ? 'text-black' : 'text-white'}`} >
-              {slider.sentence1}
-              <br/>
-              {slider.sentence2}
-            </div>
-            <img src={slider.imgsrc} className="absolute right-5 md:right-8 bottom-5 md:bottom-8"/>
-          </div>
-        ))}
-      </Slider>
-  );
-};
-
 
 //메인버튼 컴포넌트
 const MainButton = ({adverb, buttonName, imgsrc}) => {
   return (
-    <div className="w-[48%] h-[10rem] border-2 rounded-xl bg-[#f4f4f4] relative">
+    <div className="w-[48%] h-[10rem] border-2 rounded-xl bg-[#f4f4f4] relative cursor-pointer">
       <div className="absolute left-4 top-4 text-xl md:text-2xl">
         <span className="text-[#5B5B5B] font-bold">{adverb}</span>
         <br/>
@@ -97,7 +51,7 @@ const MainButton = ({adverb, buttonName, imgsrc}) => {
 //치매센터 버튼 컴포넌트
 const DementiaCenterButton = () => {
   return(
-    <div className="my-7 flex justify-between items-center border-2 py-5 px-3 md:px-8 rounded-xl">
+    <div className="my-7 flex justify-between items-center border-2 py-5 px-3 md:px-8 rounded-xl cursor-pointer">
       <div>
         <span className="text-md sm:text-lg md:text-xl font-bold">가까운 치매센터 찾기 {'>'}</span>
         <br/>
@@ -111,7 +65,7 @@ const DementiaCenterButton = () => {
 //체조 버튼 컴포넌트 
 const GymnasticsButton = () => {
   return (
-    <div className="mb-6 flex justify-between items-center border-2 rounded-xl p-5 md:px-12 bg-[#f4f4f4] text-md sm:text-lg md:text-xl">
+    <div className="mb-6 flex justify-between items-center border-2 rounded-xl p-5 md:px-12 bg-[#f4f4f4] text-md sm:text-lg md:text-xl cursor-pointer">
       <img src={btn_gymnastics} />
       <div className="mr-8 md:mr-16">
         <span className="text-orange-500">간단한 체조</span><span>를 통해</span>
@@ -133,7 +87,7 @@ const GameButtonList = () => {
   return (
     <>
       {games.map((game, index) => (
-        <div key={index} className="h-[4rem] flex justify-between items-center border-2 rounded-xl px-4 mb-5 text-lg md:text-2xl">
+        <div key={index} className="h-[4rem] flex justify-between items-center border-2 rounded-xl px-4 mb-5 text-lg md:text-2xl cursor-pointer">
           <img src={game.imgsrc} alt={game.name} width={"50px"} />
           {game.name}
           <span className="text-blue-600 font-bold">도전 {'>'}</span>
